pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "firmia"
version = "0.1.0"
description = "French Company Intelligence MCP Server"
authors = [{name = "Loic", email = "loic@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.9.3",
    "httpx[http2]>=0.27",
    "pydantic>=2.7",
    "redis[hiredis]>=5.0",
    "duckdb>=0.10",
    "uvicorn[standard]>=0.30",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "prometheus-client>=0.20",
    "structlog>=24.1",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
    "boto3>=1.34",
    "aiofiles>=23.2",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "black>=24.0",
    "ruff>=0.1.9",
    "mypy>=1.8",
    "httpx-mock>=0.4",
    "aioresponses>=0.7",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "C90", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]